pregled pacijenata, pretraga, sortiranje; izmena podataka o klinici za admina klinike

+++ MRS2020_T10/src/main/java/main/mrs/service/KlinikaService.java
@@ -16,7 +16,7 @@ public class KlinikaService {
 	@Autowired
 	private KlinikaRepository KlinikaRepository;
 	
-	public Klinika findOne(Long id) {
+	public Klinika findOne(Integer id) {
 		return KlinikaRepository.findById(id).orElseGet(null);
 	}
